Redesign sidebar menu and reorganise page structure
