+++ src/pages/build-status-page/index.ts
@@ -1,0 +1,1 @@
+export { default } from './BuildStatusPage'
