+++ src/pages/tenant-capabilities/TenantCapabilitiesTab.tsx
@@ -1,6 +1,5 @@
 import LoadingSpinner from '@/components/LoadingSpinner'
 import ErrorDisplay from '@/components/ErrorDisplay'
-import PageHeader from '@/components/PageHeader'
 import { useGetUserTenantById } from '@/hooks/useTenants'
 import {
   ArrowBigUp,
@@ -15,7 +14,6 @@ import {
   ZapIcon,
 } from 'lucide-react'
 import { useEffect, useRef, useState, type ReactElement } from 'react'
-import { useParams } from 'react-router-dom'
 
 const BACKEND_API = import.meta.env.VITE_BACKEND_URI
 
@@ -215,111 +213,84 @@ const CapabilityCard = ({
   )
 }
 
-const TenantCapabilities = () => {
-  const { id } = useParams<{ id: string }>()
-  const { data: tenantData, isLoading, error } = useGetUserTenantById(id || '')
+interface TenantCapabilitiesTabProps {
+  tenantId: string
+}
 
-  if (isLoading) {
+const TenantCapabilitiesTab = ({ tenantId }: TenantCapabilitiesTabProps) => {
+  const { data: tenantData, isLoading, error } = useGetUserTenantById(tenantId)
+
+  if (isLoading)
     return (
-      <div className="container p-8 flex justify-center">
+      <div className="flex justify-center p-8">
         <LoadingSpinner />
       </div>
     )
-  }
 
-  if (error) {
-    return (
-      <div className="container">
-        <ErrorDisplay error={error} context="tenant" />
-      </div>
-    )
-  }
+  if (error) return <ErrorDisplay error={error} context="tenant" />
 
-  if (!tenantData) {
-    return (
-      <div className="container">
-        <ErrorDisplay
-          error="The tenant you are looking for does not exist or has been removed."
-          context="tenant"
-        />
-      </div>
-    )
-  }
+  if (!tenantData) return null
 
   return (
-    <div className="min-h-scrdeen">
-      <PageHeader
-        className="max-w-7xl mx-auto mb-4 pb-2"
-        title="Capabilities"
-        subtitle={
-          <>
-            Explore capabilities for tenant{' '}
-            <strong>{tenantData.info.name}</strong>
-          </>
-        }
-        navigateTo={{ label: 'Back to Tenants', to: '/tenants' }}
+    <div className="grid grid-cols-1 lg:grid-cols-2 gap-4">
+      <CapabilityCard
+        title="Availability"
+        colorClass="bg-emerald-50 text-emerald-600 border border-emerald-100"
+        description="Percentage of time a service is fully functional and accessible, based on monitored status history."
+        uiUrl={`${tenantData.metadata?.instance?.ui_url}/${tenantData.info.name}/report-ar/CORE/SERVICEGROUPS`}
+        apiUrl={`${BACKEND_API}/api/tenants/${tenantData.info.name}/results/ar`}
+        apiDoc={`${BACKEND_API}/swagger-ui/#/Admin/get_v1_admin_tenants__id__status`}
+        apiAccess={`${BACKEND_API}/oidc-client`}
+        icon={<ClockIcon />}
+        docUrl="https://argoeu.github.io/argo-monitoring/docs/reports/ar#availability"
+        stats={[{ name: 'Avg Avail', value: 98.3 }]}
       />
 
-      <div className="max-w-7xl mx-auto grid grid-cols-1 lg:grid-cols-2 gap-4">
-        <CapabilityCard
-          title="Availability"
-          colorClass="bg-emerald-50 text-emerald-600 border border-emerald-100"
-          description="Percentage of time a service is fully functional and accessible, based on monitored status history."
-          uiUrl={`${tenantData.metadata?.instance?.ui_url}/${tenantData.info.name}/report-ar/CORE/SERVICEGROUPS`}
-          apiUrl={`${BACKEND_API}/api/tenants/${tenantData.info.name}/results/ar`}
-          apiDoc={`${BACKEND_API}/swagger-ui/#/Admin/get_v1_admin_tenants__id__status`}
-          apiAccess={`${BACKEND_API}/oidc-client`}
-          icon={<ClockIcon />}
-          docUrl="https://argoeu.github.io/argo-monitoring/docs/reports/ar#availability"
-          stats={[{ name: 'Avg Avail', value: 98.3 }]}
-        />
+      <CapabilityCard
+        title="Status"
+        colorClass="bg-indigo-50 text-indigo-600 border border-indigo-100"
+        description="Real-time chronological health timelines, tracking states between OK, Warning, and Critical transitions."
+        uiUrl={`${tenantData.metadata?.instance?.ui_url}/${tenantData.info.name}/report-status/CORE/SERVICEGROUPS`}
+        apiUrl={`${BACKEND_API}/api/tenants/${tenantData.info.name}/status`}
+        apiDoc={`${BACKEND_API}/swagger-ui/#/Admin/get_v1_admin_tenants__id__status`}
+        apiAccess={`${BACKEND_API}/oidc-client`}
+        icon={<Rows4 />}
+        docUrl="https://argoeu.github.io/argo-monitoring/docs/reports/status_timelines"
+        stats={[
+          { name: 'Warnings', value: 8, colorClass: 'text-amber-500' },
+          { name: 'Critical', value: 10, colorClass: 'text-red-500' },
+        ]}
+      />
 
-        <CapabilityCard
-          title="Status"
-          colorClass="bg-indigo-50 text-indigo-600 border border-indigo-100"
-          description="Real-time chronological health timelines, tracking states between OK, Warning, and Critical transitions."
-          uiUrl={`${tenantData.metadata?.instance?.ui_url}/${tenantData.info.name}/report-status/CORE/SERVICEGROUPS`}
-          apiUrl={`${BACKEND_API}/api/tenants/${tenantData.info.name}/status`}
-          apiDoc={`${BACKEND_API}/swagger-ui/#/Admin/get_v1_admin_tenants__id__status`}
-          apiAccess={`${BACKEND_API}/oidc-client`}
-          icon={<Rows4 />}
-          docUrl="https://argoeu.github.io/argo-monitoring/docs/reports/status_timelines"
-          stats={[
-            { name: 'Warnings', value: 8, colorClass: 'text-amber-500' },
-            { name: 'Critical', value: 10, colorClass: 'text-red-500' },
-          ]}
-        />
+      <CapabilityCard
+        title="Uptime"
+        colorClass="bg-amber-50 text-amber-600 border border-amber-100"
+        description="Continuous operation score depicting service stability without registered downtime or interruptions."
+        uiUrl={`${tenantData.metadata?.instance?.ui_url}/${tenantData.info.name}/report-ar/CORE/SERVICEGROUPS`}
+        apiUrl={`${BACKEND_API}/api/tenants/${tenantData.info.name}/results/uptime`}
+        apiDoc={`${BACKEND_API}/swagger-ui/#/Admin/get_v1_admin_tenants__id__status`}
+        apiAccess={`${BACKEND_API}/oidc-client`}
+        icon={<ArrowBigUp />}
+        docUrl="https://argoeu.github.io/argo-monitoring/docs/reports/ar#availability"
+        stats={[{ name: 'Avg Uptime', value: 99.8 }]}
+      />
 
-        <CapabilityCard
-          title="Uptime"
-          colorClass="bg-amber-50 text-amber-600 border border-amber-100"
-          description="Continuous operation score depicting service stability without registered downtime or interruptions."
-          uiUrl={`${tenantData.metadata?.instance?.ui_url}/${tenantData.info.name}/report-ar/CORE/SERVICEGROUPS`}
-          apiUrl={`${BACKEND_API}/api/tenants/${tenantData.info.name}/results/uptime`}
-          apiDoc={`${BACKEND_API}/swagger-ui/#/Admin/get_v1_admin_tenants__id__status`}
-          apiAccess={`${BACKEND_API}/oidc-client`}
-          icon={<ArrowBigUp />}
-          docUrl="https://argoeu.github.io/argo-monitoring/docs/reports/ar#availability"
-          stats={[{ name: 'Avg Uptime', value: 99.8 }]}
-        />
-
-        <CapabilityCard
-          title="Performance"
-          colorClass="bg-pink-50 text-pink-600 border border-pink-100"
-          description="Metric analytics for system performance monitoring, focusing on latency and response speed."
-          uiUrl={`${tenantData.metadata?.instance?.ui_url}/${tenantData.info.name}/performances`}
-          apiUrl={`${BACKEND_API}/api/tenants/${tenantData.info.name}/performance`}
-          apiDoc={`${BACKEND_API}/swagger-ui/#/Admin/get_v1_admin_tenants__id__status`}
-          apiAccess={`${BACKEND_API}/oidc-client`}
-          icon={<ZapIcon />}
-          docUrl="https://argoeu.github.io/argo-monitoring/docs/reports/ar#availability"
-          stats={[
-            { name: 'Latency (s)', value: 3.2, colorClass: 'text-amber-500' },
-          ]}
-        />
-      </div>
+      <CapabilityCard
+        title="Performance"
+        colorClass="bg-pink-50 text-pink-600 border border-pink-100"
+        description="Metric analytics for system performance monitoring, focusing on latency and response speed."
+        uiUrl={`${tenantData.metadata?.instance?.ui_url}/${tenantData.info.name}/performances`}
+        apiUrl={`${BACKEND_API}/api/tenants/${tenantData.info.name}/performance`}
+        apiDoc={`${BACKEND_API}/swagger-ui/#/Admin/get_v1_admin_tenants__id__status`}
+        apiAccess={`${BACKEND_API}/oidc-client`}
+        icon={<ZapIcon />}
+        docUrl="https://argoeu.github.io/argo-monitoring/docs/reports/ar#availability"
+        stats={[
+          { name: 'Latency (s)', value: 3.2, colorClass: 'text-amber-500' },
+        ]}
+      />
     </div>
   )
 }
 
-export default TenantCapabilities
+export default TenantCapabilitiesTab
