+++ src/pages/tenant-capabilities/index.ts
@@ -1,0 +1,1 @@
+export { default } from './TenantCapabilities'
